Version 2.0 change list:
- migarate to Java 1.8
- migrate to Dropwizard Metrics 4.x
- make Dropwizard 4.x default
- update CP dependencies to the latest artifacts

+++ flexy-pool-core/src/test/java/com/vladmihalcea/flexypool/util/ReflectionUtilsTest.java
@@ -100,7 +100,7 @@ public class ReflectionUtilsTest extends AbstractUtilsTest<ReflectionUtils> {
     public void testInvokeSetter() {
         TestObject testObject = new TestObject();
         ReflectionUtils.invokeSetter(testObject, "version", 12);
-        assertEquals(12, ReflectionUtils.getFieldValue(testObject, "version"));
+        assertEquals(12, (int) ReflectionUtils.getFieldValue(testObject, "version"));
     }
 
     @Override
